requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.2
django-debug-toolbar==2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.10.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jieba==0.39
lxml==4.3.4
mysqlclient==1.4.2.post1
parsel==1.5.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-memcached==1.59
pytz==2019.1
queuelib==1.5.0
redis==3.2.1
requests==2.22.0
Scrapy==1.7.1
selenium==3.141.0
service-identity==18.1.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
urllib3==1.25.3
virtualenv==16.6.0
virtualenvwrapper-win==1.2.5
w3lib==1.20.0
webencodings==0.5.1
Whoosh==2.7.4
zope.interface==4.6.0
